ci: fix cypress
